refactor the test classes
