+++ kyuubi-hive-jdbc/src/test/java/org/apache/kyuubi/jdbc/hive/TestJdbcDriver.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hive.jdbc;
+package org.apache.kyuubi.jdbc.hive;
 
 import org.junit.AfterClass;
 import org.junit.Assert;
@@ -86,7 +86,8 @@ public class TestJdbcDriver {
       bw = new BufferedWriter(new FileWriter(file));
       bw.write(input);
       bw.flush();
-      assertEquals(Arrays.asList(expected.split(",")), HiveConnection.parseInitFile(file.toString()));
+      assertEquals(Arrays.asList(expected.split(",")),
+        KyuubiConnection.parseInitFile(file.toString()));
     } catch(Exception e) {
       Assert.fail("Test was failed due to " + e);
     } finally {
